Added command-line parameters to override the default x-width and y-width settings.
Some documents don't have the default proportions of 1872x1404 pixels (for
example, US Letter documents). Exporting the lines file as SVGs will
result in lines placed at incorrect locations when superimposed onto the
original PDF in those cases. The problem is solved when overriding these
default settings with the proper pixel values for the document size
(1805x1404 in the case of US Letter). The proposed modification allows
users to specify these values as (optional) command line arguments. The
default behavior of the program is left unchanged.